v13/ATLAS/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyJWT>=2.8.0
python-multipart
mnemonic>=0.20
python-json-logger>=2.0.7
